feat: implement FacebookEventLink component for safe Facebook link handling
