+++ src/app/(public)/events/_components/EventRegistrationCard.tsx
@@ -1,15 +1,15 @@
 "use client";
 
 import { motion } from "framer-motion";
-import { Banknote, ExternalLink } from "lucide-react";
+import { Banknote } from "lucide-react";
 import type { Route } from "next";
 import Link from "next/link";
 import { useParams } from "next/navigation";
-import { FacebookIcon } from "@/components/icons/SocialIcons";
 import { Card, CardContent } from "@/components/ui/card";
 import { fadeInUp } from "@/lib/animations/fade";
 import { staggerContainer } from "@/lib/animations/stagger";
 import type { Tables } from "@/lib/supabase/db.types";
+import { FacebookEventLink } from "./FacebookEventLink";
 
 type Event = Tables<"Event">;
 
@@ -21,25 +21,6 @@ export default function EventRegistrationCard({
   event,
 }: EventRegistrationCardProps) {
   const { eventId } = useParams<{ eventId: string }>();
-  const safeFacebookLink = (() => {
-    const rawLink = event.facebookLink?.trim();
-    if (!rawLink) {
-      return null;
-    }
-
-    try {
-      const parsed = new URL(rawLink);
-      const isHttpProtocol =
-        parsed.protocol === "https:" || parsed.protocol === "http:";
-      if (!isHttpProtocol) {
-        return null;
-      }
-
-      return parsed.toString();
-    } catch {
-      return null;
-    }
-  })();
 
   return (
     <motion.div
@@ -79,20 +60,7 @@ export default function EventRegistrationCard({
                 Register for This Event
               </Link>
 
-              {safeFacebookLink && (
-                <a
-                  className="flex items-center gap-3 rounded-2xl border border-border bg-card px-4 py-3 text-foreground transition-colors hover:border-primary/30 hover:bg-accent"
-                  href={safeFacebookLink}
-                  rel="noopener noreferrer"
-                  target="_blank"
-                >
-                  <FacebookIcon className="h-5 w-5 text-[#1877F2]" />
-                  <span className="flex-1 font-medium">
-                    Event Facebook Link
-                  </span>
-                  <ExternalLink className="h-4 w-4 shrink-0 text-muted-foreground" />
-                </a>
-              )}
+              <FacebookEventLink url={event.facebookLink} />
             </div>
           </CardContent>
         </Card>
